522: v1.11: add get_csv and get_zip, modified check_list: remove unuse debug info
